requirements.txt
beautifulsoup4>=4.9.3
boto3>=1.16.9
botocore>=1.19.9
feather-format>=0.4.1
fsspec>=2024.3.1
gensim>=3.8.3
lxml>=5.0.2
nltk>=3.6.2
numpy>=1.16.6
openai==1.55.3
pandas>=0.24.2
pyarrow>=0.16.0
pyyaml>=5.4.1
requests>=2.24.0
s3fs>=2024.3.1
s3transfer>=0.3.3
scikit-learn>=0.20.4
scipy<1.13
urllib3>=1.25.11
